make logger more compact + shader loading
