+++ resources/shaders/fragment_shader.frag
@@ -1,4 +1,4 @@
-#version 140
+#version 150
 
 in vec3 v_normal;
 out vec4 color;
