--- src/info_types.rs
@@ -1,16 +1,0 @@
-#[allow(dead_code)]
-pub enum InfoTypes {
-    INFO,
-    ERROR,
-    WARNING
-}
-
-impl InfoTypes {
-    pub fn types(&self) -> &str {
-        match self {
-            InfoTypes::INFO => "Info",
-            InfoTypes::ERROR => "Error",
-            InfoTypes::WARNING => "Info",
-        }
-    }
-}
