+++ src/logger.rs
@@ -1,8 +1,27 @@
-use chrono::{Local, Timelike};
+#[allow(dead_code)]
+pub enum InfoTypes {
+    INFO,
+    ERROR,
+    WARNING
+}
 
-pub fn log(message: &str, info_type: &str) {
-    let hours = Local::now().hour();
-    let minutes = Local::now().minute();
-    let seconds = Local::now().second();
-    println!("<{}:{}:{}|{}> {}", hours, minutes, seconds, info_type, message);
+impl InfoTypes {
+    pub fn literal(&self) -> &str {
+        match self {
+            InfoTypes::INFO => "Info",
+            InfoTypes::ERROR => "Error",
+            InfoTypes::WARNING => "Info",
+        }
+    }
+}
+
+#[macro_export]
+macro_rules! log {
+    ($message: expr, $info_type: path) => {
+        let (hours, minutes, seconds) = get_current_time();
+        println!("<{}:{}:{}|{}> {}", hours, minutes, seconds, $info_type.literal(), $message);
+    };
+    ($message: expr) => {
+        log!($message, $crate::logger::InfoTypes::INFO)
+    };
 }
