+++ src/util.rs
@@ -1,12 +1,12 @@
 use std::io::Cursor;
 use glium::texture::SrgbTexture2d;
+use chrono::{Local, Timelike};
 
-use crate::logger::log;
-use crate::info_types::InfoTypes::*;
+use crate::logger::InfoTypes::*;
 
 pub fn load_image(image_bytes: &[u8], image_name: &str, display: &glium::Display) -> SrgbTexture2d {
     let logger_path = format!("Loaded image: {}", image_name);
-    log(&logger_path, INFO.types());
+    log!(&logger_path, INFO);
 
     // Load the image from the byte slice
     let image = image::load(Cursor::new(image_bytes), image::ImageFormat::Png)
@@ -19,4 +19,15 @@ pub fn load_image(image_bytes: &[u8], image_name: &str, display: &glium::Display
     // Convert to texture
     let texture = glium::texture::SrgbTexture2d::new(display, image).unwrap();
     texture
+}
+
+pub fn load_shader(file_bytes: &str, file_name: &str) -> String {
+    let logger_path = format!("Loaded shader: {}", file_name);
+    log!(&logger_path);
+    file_bytes.to_string()
+}
+
+pub fn get_current_time() -> (u32, u32, u32) {
+    let now = Local::now();
+    (now.hour(), now.minute(), now.second())
 }
